recursive implememntation of flood fill works well
changed to a bool array to store pixels viewd and started a new thread in which i define a bigger stack.

As well i added a stopwatch to tell the elapsed time of execution and as of now a run with scale of x10 takes 0.89sec on my new Lenovo.